pyproject.toml
[project]
name = "agentic-translation-qa"
version = "0.1.0"
description = "Agentic Translation QA with FastAPI, LangGraph, and local models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic-settings>=2.3.0",
  "langgraph>=0.2.0",
  "langchain>=0.3.0",
  "langchain-ollama>=0.2.0",
  "transformers>=4.44.0",
  "torch>=2.3.0",
  "sentence-transformers>=3.0.0",
  "faiss-cpu>=1.8.0",
  "python-Levenshtein>=0.25.0",
  "numpy>=1.26.0",
  "orjson>=3.10.0",
  "sentencepiece>=0.2.1",
  "langsmith>=0.7.36",
  "pandas>=3.0.2",
  "matplotlib>=3.10.9",
  "seaborn>=0.13.2",
  "koreanize-matplotlib>=0.1.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
